clients/requirements.txt
jsonrpcclient
requests
pyyaml
snet-cli
opencv-contrib-python